+++ CHANGELOG.md
@@ -6,6 +6,8 @@ most recent changes however.
 
 ## April
 
+- Rye's `pin` command now updates the pyproject.toml requires-python.
+
 - Rye's `install` command now accepts a `--include-dep` parameter to include
   scripts from one or more given dependencies.